Remove si-ps-conv-fastcorrs services from si-ps-fastcorrs stack.
The machine-applications version of Fast Corrector Conversion IOC [1] is
being replaced by a new implementation [2], whose deployment is tracked
in the docker-stacks repository. Remove its services from
si-ps-fastcorrs so it isn't deployed with service-start-si-ps-fastcorrs.
The individual services will be kept for now to simplify rollbacks, but
can be removed in the future if the new implementation proves itself
reliable.

[1] https://github.com/lnls-sirius/machine-applications/tree/master/si-ps-conv-fastcorrs
[2] https://github.com/cnpem/conv-fastcorrs-epics-ioc/